requirements.txt
# Core
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0

# LangChain & LangGraph
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# MCP Client
mcp>=1.0.0
httpx>=0.27.0
httpx-sse>=0.4.0

# Database
motor>=3.3.0
pymongo>=4.6.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# WebSocket
websockets>=12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
